requirements/common.txt
colorama
cryptography
Flask
Flask-Cors
Flask-JWT-Extended
Flask-Migrate
Flask-SQLAlchemy
flask-restx
minio
PyPika
python-dotenv
requests
SQLAlchemy